Cleanup
Remove toggle_action_entries empty array which is invalid C.
Remove incorrect semicolon from G_BEGIN_DECLS/G_END_DECLS macros.
Properly chain up the constructed virtual method in XfceMixer.
Use enums instead of macros for columns which show up in debuggers.
Use G_DEFINE_TYPE macro.
Renamed XfceMixer to XfceMixerContainer to solve namespace clash with
libxfce4mixer.
Use correct type for the handler id of
g_signal_handler_connect/g_signal_handler_disconnect in
xfce_mixer_bus_connect/xfce_mixer_bus_disconnect.
Rename _xfce_mixer_bus_message parameter in order not to shadow a global
variable.

+++ panel-plugin/xfce-mixer-plugin.c
@@ -154,7 +154,7 @@ struct _XfceMixerPlugin
   gboolean         ignore_bus_messages;
 
   /* GstBus connection id */
-  guint            message_handler_id;
+  gulong           message_handler_id;
 };
